+++ src/components/stages/header/Search.jsx
@@ -10,7 +10,7 @@ const Search = ({ activeNavigation, searchButtonLabel, searchInformationItems, s
     return <div className={'searchContainer'}>
         <input id={'searchPattern'} type={'text'} className={'searchContainer__inputField'}
                placeholder={searchTextPlaceholder} onKeyUp={() => search() } />
-        <input type={'button'} className={'button'} className={'searchContainer__button'}
+        <input type={'button'} className={'button searchContainer__button'}
                value={searchButtonLabel} onClick={() => search()} />
     </div>;
 }