+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 2020-11-08
+- Added   | container and component for user settings
+- Added   | setting "display language"
+- Added   | basic styling for settings
+
 ### 2020-10-30
 - Changed | item wishlist-icon is now a separate container/component
 - Changed | item link bar now a separate container/component